pc/requirements.txt
numpy
flask
pillow